backend/pyproject.toml
[tool.poetry]
name = "ecommerce-optimizer-backend"
version = "0.1.0"
description = "AI-powered e-commerce optimization platform backend"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
psycopg2-binary = "^2.9.9"
celery = {extras = ["redis"], version = "^5.3.0"}
redis = "^5.0.0"
pydantic = "^2.5.0"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"
django-environ = "^0.11.2"
beautifulsoup4 = "^4.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"
django-stubs = "^4.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

